% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{G1,numeric-method}
\alias{G1,numeric-method}
\title{Automatic description: G1,numeric,ANY-method}
\usage{
\S4method{G1}{numeric}(a1, a2)
}
\arguments{
\item{a1}{object of class:\code{numeric}, no manual documentation}

\item{a2}{no manual documentation}
}
\description{
Automatic description: G1,numeric,ANY-method
}
