% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{G1}
\alias{G1}
\title{S4 generic: G1}
\usage{
G1(a1, a2)
}
\arguments{
\item{a1}{see method arguments}

\item{a2}{see method arguments}
}
\description{
S4 generic: G1
}
\section{S4 methods for this generic}{
\itemize{
  \item \code{\link{G1,numeric-method}}
}
}

